+++ packages/native/authentication/lib/src/platform/native_auth.ios.dart
@@ -66,7 +66,7 @@ final class NativeAuthenticationIos extends NativeAuthenticationPlatform {
     final completion = Completer<Uri>();
     final completionHandler =
         ObjCBlock_ffiVoid_NSURL_NSError.listener((url, error) {
-      final uri = switch (url?.absoluteString?.toString()) {
+      final uri = switch (url?.absoluteString?.toDartString()) {
         final url? => Uri.tryParse(url),
         _ => null,
       };