+++ packages/native/storage/lib/src/local/local_storage_platform.vm.dart
@@ -35,7 +35,6 @@ abstract base class NativeLocalStoragePlatform extends NativeStorageBase
 
   @protected
   NativeLocalStoragePlatform.base({
-    required super.namespace,
     this.scope,
   });