+++ packages/native/storage/lib/src/native_memory_storage.dart
@@ -23,7 +23,7 @@ final class NativeMemoryStorage extends NativeStorageBase
   }
 
   NativeMemoryStorage._({
-    super.namespace,
+    String? namespace,
     this.scope,
     Map<String, String>? storage,
   })  : namespace = namespace ?? 'default',